requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
cryptography==44.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jeepney==0.8.0
pycparser==2.22
python-telegram-bot==21.7
SecretStorage==3.3.3
sniffio==1.3.1
yt-dlp==2024.12.3